Added tables.py for the tables API code.
